rename to generic to add program-specific file
