+++ harpy/deconvolve.py
@@ -6,7 +6,7 @@ import yaml
 from rich import box
 from rich.table import Table
 import rich_click as click
-from ._cli_types import HPCProfile, SnakemakeParams
+from ._cli_types_generic import HPCProfile, SnakemakeParams
 from ._conda import create_conda_recipes
 from ._launch import launch_snakemake
 from ._misc import fetch_rule, snakemake_log
