+++ harpy/phase.py
@@ -9,7 +9,7 @@ import rich_click as click
 from ._conda import create_conda_recipes
 from ._launch import launch_snakemake
 from ._misc import fetch_rule, fetch_report, snakemake_log
-from ._cli_types import ContigList, HPCProfile, InputFile, SnakemakeParams
+from ._cli_types_generic import ContigList, HPCProfile, InputFile, SnakemakeParams
 from ._parsers import parse_alignment_inputs
 from ._validations import check_fasta, vcf_sample_match, validate_bam_RG, vcf_contig_match
